+++ src/components/DownloadMenu/index.js
@@ -1,2 +1,2 @@
-export { ToolbarDownloadDropdown } from './ToolbarDownloadDropdown.js'
-export { ModalDownloadDropdown } from './ModalDownloadDropdown.js'
+export { ToolbarDownloadDropdown } from './ToolbarDownloadDropdown.jsx'
+export { ModalDownloadDropdown } from './ModalDownloadDropdown.jsx'