+++ src/components/VisualizationOptions/Options/RowSubTotals.jsx
@@ -1,0 +1,14 @@
+import i18n from '@dhis2/d2-i18n'
+import React from 'react'
+import CheckboxBaseOption from './CheckboxBaseOption.jsx'
+
+const RowSubTotals = () => (
+    <CheckboxBaseOption
+        label={i18n.t('Row sub-totals')}
+        option={{
+            name: 'rowSubTotals',
+        }}
+    />
+)
+
+export default RowSubTotals